+++ src/components/auth/index.js
@@ -1,0 +1,5 @@
+import Form from './FormComponent';
+
+export default {
+  Form
+};
